third_party/WebKit/LayoutTests/http/tests/devtools/load-file-resource-for-frontend.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script>
function test() {
  function testLoadForURL(url, next) {
    var documentURL = TestRunner.resourceTreeModel.mainFrame.url;
    var lastIndexOfSlash = documentURL.lastIndexOf('/');
    var urlPrefix = documentURL.substr(0, lastIndexOfSlash + 1);
    var fullURL = urlPrefix + url;
    TestRunner.addResult('Loading resource from ' + url);

    Host.ResourceLoader.load(fullURL, null, callback);

    function callback(statusCode, headers, content) {
      if (!content)
        TestRunner.addResult('Failed: ' + statusCode);
      else {
        TestRunner.addResult('Content: ' + content);
      }
      TestRunner.deprecatedRunAfterPendingDispatches(next);
    }
  }

  TestRunner.runTestSuite([
    function testResourceFromFileScheme(next) {
      testLoadForURL('resources/file-resource-to-load-for-frontend.txt', next);
    },
  ]);
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
