third_party/WebKit/LayoutTests/http/tests/devtools/load-file-resource-for-frontend-expected.txt


Running: testResourceFromFileScheme
Loading resource from resources/file-resource-to-load-for-frontend.txt
Content: SUCCESS
